backend/pyproject.toml
[tool.isort]
profile = "black"

[tool.poetry]
name = "vengeful-vineyard"
version = "0.1.0"
description = ""
authors = ["Dotkom"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.9.0"
uvicorn = "^0.16.0"
asgi-lifespan = "^1.0.1"
asyncpg = "^0.26.0"
aioresponses = "^0.7.3"
pytz = "^2022.6"
types-pytz = "^2022.6.0.1"
fastapi = "^0.86.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
mypy = "^0.961"
pre-commit = "^2.20.0"
pylint = "^2.14.4"
mypy-extensions = "^0.4.3"
httpx = "^0.23.0"
pytest-asyncio = "^0.18.3"
aiohttp = "^3.8.1"
pdoc3 = "^0.10.0"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
